feat: Allow users to copy auth code when automation is unavailable
We discovered it isn't always possible to automatically copy the
auth code due to the way JupyterHub mounts its urls. Since the user
changes as part of the URL, a static redirect_uri won't work. For Example:

https://my-hub.com/user/nick/lab

where `nick` is a dynamic URL for each user that starts a single
user server. This results in a different redirect_uri for each user.

The changes here allow a manual code copy, which fixes this problem.