feat: Support Mapped collections via re-login with data_access scope
These changes enable discovering the scopes requried to access an
endpoint which require an extra GCS data_access scope. This isn't always
a dependent scope on the transfer scope, since we want to allow custom
submissions in this case, so the extra GET endpoint will automatically
determine the correct 'base' scope to use and return a valid login scope
including the GCS data_access dependency.