fix: improper 400 returned by endpoint_search
When a 401, 403 was returned in a Globus Response, the transfer
exception handling would attempt to fetch an 'endpoint' query param
in order to check if a data_access scope needed to be added to the
list of scopes for a 'next' login. On operations like endpoint_search
which do not have an 'endpoint' query param, this resulted in an
improper 400.