fix: Regression in custom transfer submission not picking up scope
A special function was needed in the login_manager to lookup a special
custom transfer submission scope for fetching the access token for transfer.
Since this was the only piece of code that needs this and it wasn't previously
covered by unit tests, I deleted it thinking it was dead code. Unit tests now
cover this, and ensure it both exists and works properly.