fix: Mapped collections not working when using a custom resource server
When a custom resource server was configured with a custom scope,
JupyterLab wasn't properly requesting dependent scopes to allow transfers
to take place on the custom collections.

The scopes are now fixed. With the fix applied, users will need to logout
then login again in order for transfers to work properly.